pyproject.toml
[tool.poetry]
name = "prompt-evaluator"
version = "0.1.0"
description = "LLM 프롬프트 정량 평가 엔진"
readme = "README.md"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "prompt_evaluator"}]

[tool.poetry.dependencies]
python = "^3.10"
langsmith = "^0.3.0"
langchain = "^0.3.0"
langchain-openai = "^0.3.0"
langchain-google-vertexai = "^2.0.0"
openai = "^1.0.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.0"
typer = "^0.21.1"
rich = "^13.0.0"
langfuse = "^3.12.0"
langchain-anthropic = "^0.3.22"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
ruff = "^0.9.0"

[tool.poetry.scripts]
prompt-eval = "prompt_evaluator.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
